Fix security issues and code quality improvements
- Fix integer overflow in PrivacyAirdropPLONK.sol and PrivacyAirdrop.sol
  Changed  to  to prevent overflow

- Fix PLONK proof validation to be less restrictive
  Removed overly restrictive check that rejects proof elements with value 1

- Fix proof element parsing in relayer
  Changed from  to  for hex strings

- Reduce MAX_CLAIM_DEADLINE from 5 years to 1 year
  More reasonable deadline for claiming tokens

- Fix sanitize_nullifier logic for short nullifiers
  Improved handling when nullifier length is 16 or less

- Remove unused cli.proof_system argument
  Cleaned up unused command-line argument

+++ contracts/src/PrivacyAirdrop.sol
@@ -84,7 +84,7 @@ contract PrivacyAirdrop is BasePrivacyAirdrop {
     ) external nonReentrant validClaim(recipient, nullifier) {
         uint256[3] memory publicSignals;
         publicSignals[0] = uint256(MERKLE_ROOT);
-        publicSignals[1] = uint160(recipient);
+        publicSignals[1] = uint256(uint160(recipient));
         publicSignals[2] = uint256(nullifier);
 
         require(VERIFIER.verifyProof(proof.a, proof.b, proof.c, publicSignals), "Invalid proof");
